srcs/django/requirements.txt
django
psycopg
djangorestframework
channels
daphne
channels-redis
djangorestframework-simplejwt
adrf
redgreenunittest